refactor(onboarding): revert cosmetic move of steps/widgets to original paths
