+++ apps/mobile_app/lib/features/catalog/presentation/pages/catalog_page.dart
