+++ apps/mobile_app/lib/features/catalog/presentation/pages/learn_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mobile_app/core/models/track.dart';
 import 'package:mobile_app/features/catalog/presentation/viewmodels/learn_view_model.dart';
-import 'package:mobile_app/features/catalog/widgets/track_card.dart';
+import 'package:mobile_app/features/catalog/presentation/widgets/track_card.dart';
 
 class LearnPage extends ConsumerWidget {
   const LearnPage({super.key});
