+++ apps/mobile_app/lib/features/shell/presentation/pages/app_shell.dart
